Added dust particles and slow effect when off road, centered camera, adjusted curve speed
Added the dust particles to both tires so that when the player goes off track the car slows down a bit. Also centered the camera at the start and adjusted the speed at which different types of segments come and the amount of turn you need to make to not go off track.